don't cast -1 to unsigned long

+++ mongovi.c
@@ -408,14 +408,14 @@ int exec_count(mongoc_collection_t *collection)
 // parse update command, expect two json objects, a selector, and an update doc and exec
 int exec_update(mongoc_collection_t *collection, const char *line)
 {
-  size_t offset;
+  long offset;
   char query_doc[MAXDOC];
   char update_doc[MAXDOC];
   bson_error_t error;
   bson_t query, update;
 
   // read first json object
-  if ((offset = relaxed_to_strict(query_doc, MAXDOC, line, strlen(line), 1)) == (size_t)-1)
+  if ((offset = relaxed_to_strict(query_doc, MAXDOC, line, strlen(line), 1)) == -1)
     return ILLEGAL;
   if (offset == 0)
     return ILLEGAL;
@@ -424,7 +424,7 @@ int exec_update(mongoc_collection_t *collection, const char *line)
   line += offset;
 
   // read second json object
-  if ((offset = relaxed_to_strict(update_doc, MAXDOC, line, strlen(line), 1)) == (size_t)-1)
+  if ((offset = relaxed_to_strict(update_doc, MAXDOC, line, strlen(line), 1)) == -1)
     return ILLEGAL;
   if (offset == 0)
     return ILLEGAL;
@@ -456,13 +456,13 @@ int exec_update(mongoc_collection_t *collection, const char *line)
 // parse insert command, expect one json objects, the insert doc and exec
 int exec_insert(mongoc_collection_t *collection, const char *line, int len)
 {
-  size_t offset;
+  long offset;
   char insert_doc[MAXDOC];
   bson_error_t error;
   bson_t doc;
 
   // read first json object
-  if ((offset = relaxed_to_strict(insert_doc, MAXDOC, line, len, 1)) == (size_t)-1)
+  if ((offset = relaxed_to_strict(insert_doc, MAXDOC, line, len, 1)) == -1)
     return ILLEGAL;
   if (offset == 0)
     return ILLEGAL;
@@ -494,7 +494,7 @@ int exec_query(mongoc_collection_t *collection, const char *line, int len)
   char query_doc[MAXDOC];
 
   // try to parse as relaxed json and convert to strict json
-  if (relaxed_to_strict(query_doc, MAXDOC, line, len, 0) == (size_t)-1)
+  if (relaxed_to_strict(query_doc, MAXDOC, line, len, 0) == -1)
     fatal("jsonify error");
 
   // try to parse it as json and convert to bson
@@ -534,7 +534,7 @@ int exec_agquery(mongoc_collection_t *collection, const char *line, int len)
   char query_doc[MAXDOC];
 
   // try to parse as relaxed json and convert to strict json
-  if (relaxed_to_strict(query_doc, MAXDOC, line, len, 0) == (size_t)-1)
+  if (relaxed_to_strict(query_doc, MAXDOC, line, len, 0) == -1)
     fatal("jsonify error");
 
   // try to parse it as json and convert to bson
